Change one assert to test+throw
The caller is not limited in its choice of timestamps, and if it
is not using the aggregator for all ports, it may actually end up
asking for a future timestamp, like it could for a past timestamp.
Instead of aborting the app, output an error message and throw an
exception.